requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
httpx==0.28.1
python-dotenv==1.2.1
python-dateutil==2.9.0
openai==2.11.0
google-cloud-speech==2.34.0
pytest==9.0.2
